# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Quantum Commerce API
  description: API oficial do case fictício Quantum Commerce — disciplina Tools, Automations and Workflows (FIAP). Backend
    determinístico para agentes n8n, Dify, Power Automate e Python. Sem LLM no backend.
  version: 2.0.0
paths:
...
  /api/v1/orders/{order_id}:
    get:
      tags:
      - Orders
      summary: Get Order By Id
      operationId: get_order_by_id
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          title: Order Id
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/{order_id}/shipment:
    get:
      tags:
      - Orders
      summary: Get Order Shipment
      operationId: get_order_shipment
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          title: Order Id
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentResponse'
        '404':
          description: Order or shipment not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/shipments/{shipment_id}:
    get:
      tags:
      - Logistics
      summary: Get Shipment By Id
      operationId: get_shipment_by_id
      parameters:
      - name: shipment_id
        in: path
        required: true
        schema:
          type: string
          title: Shipment Id
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentResponse'
        '404':
          description: Shipment not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/policies:
    get:
      tags:
      - Policies
      summary: List Policies Route
      operationId: list_policies
      parameters:
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
      - name: country
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Country
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Policies
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/policies/search:
    get:
      tags:
      - Policies
      summary: Search Policies Route
      operationId: search_policies
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          title: Q
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Search Policies
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/policies/{policy_id}:
    get:
      tags:
      - Policies
      summary: Get Policy Route
      operationId: get_policy
      parameters:
      - name: policy_id
        in: path
        required: true
        schema:
          type: string
          title: Policy Id
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Policy
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/meta:
    get:
      tags:
      - Meta
      summary: Get Meta
      operationId: get_meta
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Meta
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lab/slow:
    get:
      tags:
      - Lab
      summary: Lab Slow
      operationId: lab_slow
      parameters:
      - name: seconds
        in: query
        required: false
        schema:
          type: integer
          maximum: 15
          minimum: 1
          default: 5
          title: Seconds
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Request-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Lab Slow
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lab/error:
    get:
      tags:
      - Lab
      summary: Lab Error
      operationId: lab_error
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Request-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Lab Error
        '500':
          description: Simulated internal server error
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lab/rate-limit:
    get:
      tags:
      - Lab
      summary: Lab Rate Limit
      operationId: lab_rate_limit
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Request-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Lab Rate Limit
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lab/not-found:
    get:
      tags:
      - Lab
      summary: Lab Not Found
      operationId: lab_not_found
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Request-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Lab Not Found
        '404':
          description: Resource not found (simulated)
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lab/validation:
    post:
      tags:
      - Lab
      summary: Lab Validation
      operationId: lab_validation
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Request-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationPayload'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Lab Validation
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OrderItemResponse:
      properties:
        sku:
          type: string
          title: Sku
        name:
          type: string
          title: Name
        quantity:
          type: integer
          title: Quantity
        unit_price:
          type: number
          title: Unit Price
      type: object
      required:
      - sku
      - name
      - quantity
      - unit_price
      title: OrderItemResponse
    OrderResponse:
      properties:
        id:
          type: string
          title: Id
        customer_id:
          type: string
          title: Customer Id
        status:
          type: string
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
        total:
          type: number
          title: Total
        currency:
          type: string
          title: Currency
        items:
          items:
            $ref: '#/components/schemas/OrderItemResponse'
          type: array
          title: Items
      type: object
      required:
      - id
      - customer_id
      - status
      - created_at
      - total
      - currency
      - items
      title: OrderResponse
...
    ShipmentEvent:
      properties:
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        location:
          type: string
          title: Location
        status:
          type: string
          title: Status
        description:
          type: string
          title: Description
      type: object
      required:
      - timestamp
      - location
      - status
      - description
      title: ShipmentEvent
    ShipmentResponse:
      properties:
        order_id:
          type: string
          title: Order Id
        shipment_id:
          type: string
          title: Shipment Id
        carrier:
          type: string
          title: Carrier
        status:
          type: string
          title: Status
        tracking_code:
          type: string
          title: Tracking Code
        estimated_delivery:
          type: string
          format: date
          title: Estimated Delivery
        last_update:
          type: string
          format: date-time
          title: Last Update
        events:
          items:
            $ref: '#/components/schemas/ShipmentEvent'
          type: array
          title: Events
      type: object
      required:
      - order_id
      - shipment_id
      - carrier
      - status
      - tracking_code
      - estimated_delivery
      - last_update
      title: ShipmentResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ValidationPayload:
      properties:
        email:
          type: string
          format: email
          title: Email
        amount:
          type: number
          minimum: 0.0
          title: Amount
      type: object
      required:
      - email
      - amount
      title: ValidationPayload
      examples:
      - amount: 42.5
        email: buyer@quantum.example
...